Move sources/ folder
