+++ packages/server/scripts/fetch-price-history.ts
@@ -1,9 +1,7 @@
 import { NETWORKS } from "@anthem/utils";
 import axios from "axios";
 import fs from "fs";
-import FIAT_CURRENCIES, {
-  FiatCurrency,
-} from "../src/server/sources/fiat-currencies";
+import FIAT_CURRENCIES, { FiatCurrency } from "../src/sources/fiat-currencies";
 
 /** ===========================================================================
  * Coin Price History Script
