+++ packages/server/src/resolvers/cosmos-resolvers.ts
@@ -23,9 +23,9 @@ import {
   IQuery,
   NETWORK_NAME,
 } from "@anthem/utils";
-import COSMOS_EXTRACTOR from "../server/sources/cosmos-extractor";
-import COSMOS_SDK from "../server/sources/cosmos-sdk";
-import EXCHANGE_DATA_API from "../server/sources/fiat-price-data";
+import COSMOS_EXTRACTOR from "../sources/cosmos-extractor";
+import COSMOS_SDK from "../sources/cosmos-sdk";
+import EXCHANGE_DATA_API from "../sources/fiat-price-data";
 import {
   blockUnsupportedNetworks,
   standardizeTimestamps,
