+++ packages/server/src/sources/cosmos-extractor.ts
@@ -9,15 +9,15 @@ import {
   NetworkDefinition,
 } from "@anthem/utils";
 import { Pool } from "pg";
-import { PaginationParams } from "../../resolvers/resolvers";
-import ENV from "../../tools/server-env";
+import { PaginationParams } from "../resolvers/resolvers";
+import ENV from "../tools/server-env";
 import {
   filterSanityCheckHeights,
   formatTransactionResponse,
   gatherEndOfDayBalanceValues,
   mapSumToBalance,
-} from "../../tools/server-utils";
-import { getSqlQueryString, SQLVariables } from "../../tools/sql-utils";
+} from "../tools/server-utils";
+import { getSqlQueryString, SQLVariables } from "../tools/sql-utils";
 
 /** ===========================================================================
  * Postgres Pool
