+++ packages/server/src/sources/fiat-currencies.ts
