+++ packages/server/src/sources/oasis.ts
@@ -6,9 +6,9 @@ import {
   IQuery,
   NetworkDefinition,
 } from "@anthem/utils";
-import { PaginationParams } from "../../resolvers/resolvers";
-import { AxiosUtil, getHostFromNetworkName } from "../../tools/axios-utils";
-import { logSentryMessage } from "../../tools/server-utils";
+import { PaginationParams } from "../resolvers/resolvers";
+import { AxiosUtil, getHostFromNetworkName } from "../tools/axios-utils";
+import { logSentryMessage } from "../tools/server-utils";
 
 /** ===========================================================================
  * Types & Config
