+++ packages/server/src/tools/server-utils.ts
@@ -13,7 +13,7 @@ import { NextFunction, Request, Response } from "express";
 import { GraphQLError } from "graphql";
 import moment from "moment-timezone";
 import { identity, ifElse, lensProp, over } from "ramda";
-import { Price } from "../server/sources/fiat-price-data";
+import { Price } from "../sources/fiat-price-data";
 import ENV from "./server-env";
 
 /** ===========================================================================
